requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
openai==1.34.0
feedparser==6.0.11
httpx==0.27.0
python-dotenv==1.0.1
transformers==4.41.1
torch>=2.0.0
slowapi==0.1.6